+++ newspack-event-logger-nodes.php
@@ -2,7 +2,7 @@
 /**
  * Plugin Name: Newspack Event Logger Nodes
  * Description: Event-logger application built on newspack-nodes runtime.
- * Version: 0.99.2
+ * Version: 0.99.3
  * Author: Automattic
  * Author URI: https://newspack.com/
  * License: GPL-2.0-or-later
@@ -40,7 +40,7 @@
 \defined( 'ABSPATH' ) || exit;
 
 if ( ! \defined( 'NEWSPACK_EVENT_LOGGER_NODES_VERSION' ) ) {
-	\define( 'NEWSPACK_EVENT_LOGGER_NODES_VERSION', '0.99.2' );
+	\define( 'NEWSPACK_EVENT_LOGGER_NODES_VERSION', '0.99.3' );
 }
 if ( ! \defined( 'NEWSPACK_EVENT_LOGGER_NODES_DIR' ) ) {
 	\define( 'NEWSPACK_EVENT_LOGGER_NODES_DIR', \plugin_dir_path( __FILE__ ) );
@@ -403,7 +403,7 @@ function newspack_event_logger_nodes_mount_service_cis( \Newspack_Nodes\Command_
 			'nonce'        => $nonce,
 			'restartNonce'      => $restart_nonce,
 			'tree'              => $tree,
-			'version'           => \NEWSPACK_NODES_VERSION,
+			'version'           => \NEWSPACK_EVENT_LOGGER_NODES_VERSION,
 		];
 
 		$handle = \Newspack_Nodes\Admin\Admin::enqueue_react_page(